enforce code file length
